third_party/blink/web_tests/external/wpt/css/css-text/white-space/seg-break-transformation-003.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] linebreak only
  assert_equals: expected 48 but got 55
[FAIL] spaces linebreak
  assert_equals: expected 48 but got 55
[FAIL] linebreak spaces
  assert_equals: expected 48 but got 55
[FAIL] spaces linebreak spaces
  assert_equals: expected 48 but got 55
[FAIL] multiple linebreaks
  assert_equals: expected 48 but got 55
[FAIL] multiple linebreaks + spaces
  assert_equals: expected 48 but got 55
Harness: the test ran to completion.
